Manage heart rate sessions and modified HeartRateScreen to use it with the timer.
I have not yet tested the code (mainly cuz I need to deploy the backend
and don't have the watch myself), but the changes are strictly related
to adding the HeartRateSessionManager file that deals with creating the
CSV of the heart rate data + the management of the data. Modified HeartRateScreen.kt
to use the HeartRateSessionManager to start recording the heart rate
data when the timer starts and create the CSV file when the timer stops.